refacto(folios) refactor statistics folio
