+++ start/routes.ts
@@ -32,7 +32,7 @@ router
           .group(() => {
             router.get('/', [ChildFoliosController, 'index']).use(middleware.auth())
             router.get('/cards', [MainFoliosController, 'cards']).use(middleware.auth())
-            router.get('/statistics', [FoliosController, 'statistics']).use(middleware.auth())
+            router.get('/statistics', [MainFoliosController, 'statistics']).use(middleware.auth())
             router.get('/:id', [FoliosController, 'show']).use(middleware.auth())
             router.get('/:id/cards', [FoliosController, 'childFolioCards']).use(middleware.auth())
             router.post('/', [CardFoliosController, 'store']).use(middleware.auth())
