+++ tests/functional/controllers/main_folios_controller.spec.ts
@@ -11,6 +11,7 @@ import { LegalityFactory } from '#database/factories/legality'
 import { SetFactory } from '#database/factories/set'
 import { SubtypeFactory } from '#database/factories/subtype'
 import { TypeFactory } from '#database/factories/type'
+import { DateTime } from 'luxon'
 
 test.group('Main folio controller', (group) => {
   let wardenApiClientStub: sinon.SinonStub
@@ -580,5 +581,248 @@ test.group('Main folio controller', (group) => {
     const returnedCardFolio = body.data[0]
     assert.equal(returnedCardFolio.card.id, matchingCard.id)
     assert.equal(returnedCardFolio.occurrence, 1)
+  })
+
+  test('statistics - should return folio statistics with correct structure', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const todayCards = await CardFactory.with('cardMarketPrices', 1, (cardMarketPrices) =>
+      cardMarketPrices.merge({
+        trendPrice: 10.5,
+        reverseHoloTrend: 15.75,
+        updatedAt: DateTime.now(),
+      })
+    )
+      .with('tcgPlayerReportings', 1, (tcgPlayerReportings) =>
+        tcgPlayerReportings
+          .merge({
+            updatedAt: DateTime.now(),
+          })
+          .with('tcgPlayerPrices', 2, (tcgPlayerPrices) =>
+            tcgPlayerPrices.merge([
+              { type: 'normal', market: 12.0 },
+              { type: 'holofoil', market: 18.5 },
+            ])
+          )
+      )
+      .createMany(3)
+
+    await CardFolioFactory.merge([
+      { cardId: todayCards[0].id, folioId: mainFolio.id, occurrence: 2 },
+      { cardId: todayCards[1].id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: todayCards[2].id, folioId: mainFolio.id, occurrence: 3 },
+    ]).createMany(3)
+
+    const response = await client
+      .get('/api/v1/folios/statistics')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(200)
+
+    const statistics = response.body()
+    assert.properties(statistics, [
+      'totalCardsCount',
+      'cardMarketPrice',
+      'tcgPlayerPrice',
+      'cardMarketTrending',
+      'tcgPlayerTrending',
+    ])
+
+    assert.isNumber(statistics.totalCardsCount)
+    assert.isString(statistics.cardMarketPrice)
+    assert.isString(statistics.tcgPlayerPrice)
+    assert.isString(statistics.cardMarketTrending)
+    assert.isString(statistics.tcgPlayerTrending)
+
+    assert.equal(statistics.totalCardsCount, 6) // 2 + 1 + 3 occurrences
+    assert.match(statistics.cardMarketPrice, /^\d+(\.\d{2})?$/) // Format prix: "10.50"
+    assert.match(statistics.tcgPlayerPrice, /^\d+(\.\d{2})?$/)
+    assert.oneOf(statistics.cardMarketTrending, ['up', 'down', 'equal'])
+    assert.oneOf(statistics.tcgPlayerTrending, ['up', 'down', 'equal'])
+  })
+
+  test('statistics - should handle cards without recent prices', async ({ client, assert }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const cardsWithoutPrices = await CardFactory.createMany(2)
+
+    await CardFolioFactory.merge([
+      { cardId: cardsWithoutPrices[0].id, folioId: mainFolio.id, occurrence: 1 },
+      { cardId: cardsWithoutPrices[1].id, folioId: mainFolio.id, occurrence: 2 },
+    ]).createMany(2)
+
+    const response = await client
+      .get('/api/v1/folios/statistics')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(200)
+
+    const statistics = response.body()
+    assert.equal(statistics.totalCardsCount, 3) // 1 + 2 occurrences
+    assert.equal(statistics.cardMarketPrice, '0.00')
+    assert.equal(statistics.tcgPlayerPrice, '0.00')
+    assert.equal(statistics.cardMarketTrending, 'equal')
+    assert.equal(statistics.tcgPlayerTrending, 'equal')
+  })
+
+  test('statistics - should only include cards from main folio', async ({ client, assert }) => {
+    const userId1 = TEST_AUTH_USER_ID
+    const userId2 = 'other-user-id'
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio1 = await FolioFactory.merge({
+      userId: userId1,
+      isRoot: true,
+    }).create()
+
+    const mainFolio2 = await FolioFactory.merge({
+      userId: userId2,
+      isRoot: true,
+    }).create()
+
+    const secondaryFolio1 = await FolioFactory.merge({
+      userId: userId1,
+      isRoot: false,
+    }).create()
+
+    const cards = await CardFactory.with('cardMarketPrices', 1, (cardMarketPrices) =>
+      cardMarketPrices.merge({
+        trendPrice: 10.0,
+        updatedAt: DateTime.now(),
+      })
+    ).createMany(4)
+
+    await CardFolioFactory.merge([
+      { cardId: cards[0].id, folioId: mainFolio1.id, occurrence: 1 },
+      { cardId: cards[1].id, folioId: mainFolio1.id, occurrence: 2 },
+    ]).createMany(2)
+
+    await CardFolioFactory.merge({
+      cardId: cards[2].id,
+      folioId: mainFolio2.id,
+      occurrence: 1,
+    }).create()
+
+    await CardFolioFactory.merge({
+      cardId: cards[3].id,
+      folioId: secondaryFolio1.id,
+      occurrence: 1,
+    }).create()
+
+    const response = await client
+      .get('/api/v1/folios/statistics')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(200)
+
+    const statistics = response.body()
+    assert.equal(statistics.totalCardsCount, 3) // Only cards from user1's main folio: 1 + 2
+    assert.equal(statistics.cardMarketPrice, '30.00') // 3 cards × 10.0 each
+  })
+
+  test('statistics - should return 404 when user has no main folio', async ({ client }) => {
+    const response = await client
+      .get('/api/v1/folios/statistics')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(404)
+    response.assertBodyContains({
+      code: 'E_ROW_NOT_FOUND',
+    })
+  })
+
+  test('statistics - should require authentication', async ({ client }) => {
+    const response = await client.get('/api/v1/folios/statistics')
+
+    response.assertStatus(401)
+  })
+
+  test('statistics - should handle empty folio correctly', async ({ client, assert }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const response = await client
+      .get('/api/v1/folios/statistics')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(200)
+
+    const statistics = response.body()
+    assert.equal(statistics.totalCardsCount, 0)
+    assert.equal(statistics.cardMarketPrice, '0.00')
+    assert.equal(statistics.tcgPlayerPrice, '0.00')
+    assert.equal(statistics.cardMarketTrending, 'equal')
+    assert.equal(statistics.tcgPlayerTrending, 'equal')
+  })
+
+  test('statistics - should properly aggregate card occurrences in calculations', async ({
+    client,
+    assert,
+  }) => {
+    const userId = TEST_AUTH_USER_ID
+
+    await ArtistFactory.create()
+    await RarityFactory.create()
+    await LegalityFactory.create()
+    await SetFactory.merge({ id: 'base1' }).create()
+
+    const mainFolio = await FolioFactory.merge({
+      userId,
+      isRoot: true,
+    }).create()
+
+    const card = await CardFactory.with('cardMarketPrices', 1, (cardMarketPrices) =>
+      cardMarketPrices.merge({
+        trendPrice: 10.0,
+        updatedAt: DateTime.now(),
+      })
+    ).create()
+
+    await CardFolioFactory.merge({
+      cardId: card.id,
+      folioId: mainFolio.id,
+      occurrence: 5,
+    }).create()
+
+    const response = await client
+      .get('/api/v1/folios/statistics')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(200)
+
+    const statistics = response.body()
+    assert.equal(statistics.totalCardsCount, 5) // Should count all occurrences
+    assert.equal(statistics.cardMarketPrice, '50.00') // 5 cards × 10.0 each
   })
 })
